Move some functions to input.cpp
