+++ casino.cpp
@@ -9,6 +9,7 @@
 #include "dmgheal.hpp"
 #include "elona.hpp"
 #include "i18n.hpp"
+#include "input.hpp"
 #include "item.hpp"
 #include "itemgen.hpp"
 #include "macro.hpp"
