+++ variables.hpp
@@ -938,8 +938,6 @@ void imeset(int = 0);
 void incognitobegin();
 void incognitoend();
 std::pair<int, int> inv_getheader(int);
-void key_check(int = 0);
-void keyrelease();
 void initialize_rankn();
 void initialize_post_data();
 void initialize_building_daga();
@@ -1210,7 +1208,6 @@ void msg_newline();
 void opencard(int = 0);
 void page_load();
 void page_save();
-void press(bool = false);
 void remain_make(int = 0, int = 0);
 void removeitem(int, int);
 void removeworker(int = 0);
